+++ src/Tracing/Doctrine/DBAL/TracingDriverConnection.php
@@ -166,6 +166,20 @@ final class TracingDriverConnection implements TracingDriverConnectionInterface
 
     /**
      * {@inheritdoc}
+     *
+     * @return resource|object
+     */
+    public function getNativeConnection()
+    {
+        if (!method_exists($this->decoratedConnection, 'getNativeConnection')) {
+            throw new \BadMethodCallException(sprintf('The connection "%s" does not support accessing the native connection.', \get_class($this->decoratedConnection)));
+        }
+
+        return $this->decoratedConnection->getNativeConnection();
+    }
+
+    /**
+     * {@inheritdoc}
      */
     public function errorCode(): ?string
     {