+++ src/Tracing/Doctrine/DBAL/TracingDriverConnectionInterface.php
@@ -6,6 +6,9 @@ namespace Sentry\SentryBundle\Tracing\Doctrine\DBAL;
 
 use Doctrine\DBAL\Driver\Connection;
 
+/**
+ * @method resource|object getNativeConnection()
+ */
 interface TracingDriverConnectionInterface extends Connection
 {
     public function getWrappedConnection(): Connection;