+++ tests/Tracing/Doctrine/DBAL/Fixture/ServerInfoAwareConnectionStub.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Sentry\SentryBundle\Tests\Tracing\Doctrine\DBAL;
+namespace Sentry\SentryBundle\Tests\Tracing\Doctrine\DBAL\Fixture;
 
 use Doctrine\DBAL\Driver\Connection;
 use Doctrine\DBAL\Driver\ServerInfoAwareConnection;